Revert "Refactor: mv func ALLCLOSE to module"
This reverts commit 862464463bb82024c752e9b25929cc206dcc996e.
